packages/server/requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.4  # https://github.com/redis/redis-py
hiredis==2.2.2  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.22.0  # https://github.com/encode/uvicorn
asgiref==3.6.0
channels==4.0.0
# django-eventstream==4.5.1
websocket-client==1.5.2
channels_redis==4.1.0

# Django
# ------------------------------------------------------------------------------
django==4.2  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
graphene-django==3.0.0
django-cors-headers==3.14.0
django-graphql-jwt==0.3.4
# django_filters==0.23.2
# graphene_django_filter==0.6.4

# Langchain stuff
langchain[llms]==0.0.174
